Re-enable the GitHub Actions CICD workflow
Now that I have reworked the build system for a multi-instance docs
setup and fixed all the issues that came up, it's time to put my work to
the (literal) test.

This commit re-enables the old GitHub Actions CICD workflow. However,
this time, the checks are run against:

- The live CMP public docs
- The live superQuery public docs
- My old Sphinx POC

I disabled the telemetry job because the telemetry rules still need some
work, but they can wait for later.