Remove build trigger job
Previously, I turned off auto-publishing. I thought this would disable
builds, but I was wrong.

The build trigger job works, but there is no way to prevent Netlify from
building the main branch automatically, so this job results in a
duplicate build.

The only way to properly integrate Netlify into the CI/CD pipeline is to
use the CLI tool to upload the build artifacts from the GitHub Actions
workflow.

I will return to this again in the future.