Colocate docs images with the docs
Everything in the `static` directory is copied verbatim to the build
directory. However, images referenced in the docs are processed into a
build time `assets` directory.

We never reference these images as [static-assets][static-assets], so
moving them out of the `static` directory means that Docusaurus does not
copy redundant (unused) files into the build directory.

[static-assets]: https://docusaurus.io/docs/static-assets