Add image zoom, auto-hide nav bar
This commit adds the [docusaurus-plugin-image-zoom][image-zoom] plugin,
which makes all docs images zoomable. To complement this change, I have
enabled the nav bar hide-on-scroll behavior.

A small CSS fix was required to get the zoom container to appear above
the nav bar, when visible.

At the moment, when installing the plugin, `yarn` produces a warning
about private workspaces. I have filed this as [issue #5][issue-5] on
the upstream project.

[image-zoom]: https://github.com/gabrielcsapo/docusaurus-plugin-image-zoom
[issue-5]: https://github.com/gabrielcsapo/docusaurus-plugin-image-zoom/issues/5