Configure the initial set of spotlight features
Spotlight features:

- Cloud analytics
- Rightsizing for Google Cloud
- Flexsave
- BigQuery Lens
- AWS Lens
- Spot scaling

I assembled this initial list of spotlight features to mirror the
existing docs as much as possible by:

- Cherry-picking prominent headings from the CMP docs front page and
  navigation sidebar.

- Ordering them approximately by their order of appearance in the
  aforementioned places.

The CMP docs front page and navigation sidebar are in need of some major
improvements, but that is out of scope for now.

To make the transition away from GitBook as smooth as possible, we want
to launch a Docusaurus website that closely mirrors the existing CMP
docs. After launch, we can restructure and rework what we have as we
develop out the help center.

Over time, we can use analytics (e.g., Mixpanel) and internal feedback
to improve (and maybe even automate) our selections for the spotlight
features.