Responsive images with zoom, plus refactoring
**TL;DR: Responsive images, size goes zoom zoom**

This commit adds a new `picture` component we can use in MDX for all
images in the docs. This element uses [ReactIdealImage][ideal-image] to
provide an "almost ideal image (responsive, lazy-loading, and
low-quality placeholder)." In addition, all resulting images are
zoomable at their highest resolution, thanks to integration with
[Zooming][zooming].

I managed to achieve this by using the
[docusaurus-plugin-ideal-image][docusaurus-plugin-ideal-image] for the
bulk of the functionality and configuration it provides, such as the
Webpack setup (for generating multiple image sizes for different devices
and clients). However, I had to swizzle the `Image` component and eject
the existing code so I could reimplement it from scratch.

The Docusaurus `Image` component is a light wrapper around the
`ReactIdealImage` component. However, with no way to swizzle the
(non-Docusaurus) `ReactIdealImage` component, I essentially had to mod
it by hooking into the [getMessage][getmessage] method and use it as a
proxy to intercept internal state changes, allowing me to intervene when
necessary.

The `getMessage` method was _absolutely not_ designed for this use case.

Regardless, I was able to use this callback method to listen for
interesting events (such as images about to be loaded in) so that I
could set up a mutation observer and react to any changes made to the
DOM.

Specifically, I use this mutation observer to:

- Reconfigure the UI to simplify and improve its design

- Attach and detach individual zoom listeners to the images as they load
  in and out

Originally, I tried to achieve all this by getting the
[docusaurus-plugin-image-zoom][docusaurus-plugin-image-zoom] working
with [docusaurus-plugin-ideal-image][docusaurus-plugin-ideal-image]:

- https://github.com/doitintl/help/pull/60

However, the two plugins are incompatible because:

- Images are dynamically loaded into their placeholders as they scroll
  into the viewport (vs. being loaded in all at once for every new page,
  as usual)

- The zoom plugin waits for new routes (when a user navigates to a new
  page) and then attaches to all of the images on the new page

My reimplementation of the `Image` component solves this conflict. I can
respond to every image load event and attach a zoom listener.

I decided to use the Zooming plugin over the [medium-zoom][medium-zoom]
plugin (used by docusaurus-plugin-image-zoom) because it functions more
reliably, is more featureful, is compatible with Retina resolution
(i.e., 2x), and can be configured to zoom in using the largest available
image (regardless of whatever size was chosen for the image on the
page).

This PR was a passion project over my holiday break and my first real
dive into React.

Because the implementation is, essentially, a pile of monkey patches in
a trenchcoat, I went through several complete redesigns before settling
on the one we have today. The result seems reliable and works well. It
may even be a good idea to contribute some of this work upstream or
release it as a standalone plugin.

Other notable changes:

- We standardized on Retina resolution a while back when updating the
  bulk of our screenshots for the new CMP navigation. However, many of
  the screenshots we didn't update still use a standard resolution.

  Because my implementation uses a 50% reduced `naturalWidth` and
  `naturalHeight` to compensate for the Retina source files (and there
  is no reliable way to detect Retina resolution), any non-Retina images
  would appear at 50% of their natural size.

  To fix this issue (temporarily, until we eventually replace them), I
  decided to upscale the standard resolution images using
  [waifu2x][waifu2x], which is remarkably good at improving the
  sharpness of the image (when compared to techniques that do not use
  Deep Convolutional Neural Networks).

- The `docusaurus.config.js` file was getting unwieldy, so I decomposed
  some of the larger logical chunks into smaller, more manageable files.

- Because of my insatiable desire to optimize side speed, I decided to
  shrink the file size of our source images by manually using
  [pngquant][pngquant]. Full integration into the build tools will have
  to wait.

[ideal-image]: https://github.com/stereobooster/react-ideal-image/blob/master/introduction.md
[zooming]: https://github.com/kingdido999/zooming
[docusaurus-plugin-ideal-image]: https://github.com/facebook/docusaurus/tree/main/packages/docusaurus-plugin-ideal-image
[docusaurus-plugin-image-zoom]: https://github.com/gabrielcsapo/docusaurus-plugin-image-zoom
[getmessage]: https://github.com/stereobooster/react-ideal-image#getmessage
[medium-zoom]: https://www.npmjs.com/package/medium-zoom
[waifu2x]: https://github.com/nagadomi/waifu2x
[pngquant]: https://pngquant.org/