Add lazy loading videos with progressive rendering
This commit combines the two previous components, `LoomEmbed` and
`YouTubeEmbed`, into a single theme component named `VideoEmbed`. This
component is available for use in all MDX files using the `video`
element.

Our docs use a lot of videos. Arguably, most of them are no longer
necessary. However, replacing them with the appropriate content is
outside this PR's scope.

I wanted to move quickly and improve the performance of any pages that
embed videos. Performance is a significant concern for the default
landing page of the CMP docs, which embeds EIGHT externally-hosted
videos that the browser attempts to load as soon as you visit the page.

To address this problem, I am using the [react-lazyload][react-lazyload]
plugin and wrapping the `VideoEmbed` component in a `LazyLoad`
component.

The new lazy-loading behavior means that the browser only attempts to
load video as it is scrolled into view, thereby significantly improving
page performance by eliminating unnecessary CPU cycles and network
requests.

In addition, I have enhanced the `VideoEmbed` component to support
pre-loading a video thumbnail. Both YouTube and Loom provide video
thumbnail URLs. However, for Loom videos, you must poke around in the
dev inspector to find them.

With thumbnail support, this is how videos are now loaded:

- When initially mounted, the component produces a simple div element
  with a neutral background color and an SVG loading icon

- When the component is about to be scrolled into view, the video
  thumbnail is loaded in to replace the background image

- At the same time, the video embed is initialized and eventually
  replaces the thumbnail image

[react-lazyload]: https://github.com/twobin/react-lazyload