Bring all Flexsave docs together
NOTE: This change is a proposal. I am staging it here so you can
interact with it and see how you feel. I expect my reviewers to have
comments, suggestions, requests, and so on—all fine.

I aim to present a new way of interacting with the Flexsave docs to see
if it sparks any ideas for additional improvements.

---

Previously, the bulk of our Flexsave docs (and the info customers will
need, regardless of their user journey) was located in the top-level CMP
section. Then, the new onboarding docs were split off to make a new
top-level Flexsave section.

This bifurcation left the new Flexsave section looking very bare. And
any time we want to reference info covered by the CMP docs, we would
have to either:

- Copy it (violating the DRY principle and probably upsetting Google's
  canonical page algorithm), or

- We would have to sprinkle links everywhere leading back to the CMP
  docs (which would be disorienting for readers)

In the long term, we plan to combine all of the Flexsave docs and
present them as a single product.

I figure now is as good a time as any to make that switch.

Accordingly, I have brought everything together to form a new top-level
Flexsave section containing all of our Flexsave docs.

I experimented with the sidebar structure to find one that works for the
content we have right now.

I know we agreed on "onboarding," but in this context, I think "sign up"
is the operative verb. It makes more sense in the context of the sidebar
tree structure. Onboarding is still there but as a component of the
sign-up process.

Notes:

- I added a replacement "Flexsave" link to the CMP sidebar to mimic the
  old section so that CMP users are still aware of it as a feature
  within the CMP.

  However, selecting that nav item takes you to the new top-level
  Flexsave section. It is clear from the UI that you have jumped to an
  entirely new section (and can return if you want to).

- I know we previously agreed on a set of URLs for the engineers to use
  in the CMP. We can continue to use those URLs with the power of HTTP
  redirects.

  However, if there is still time, I can generate a new list of URLs for
  the engineers.

- I added a placeholder page when you select the "Sign up" link from the
  sidebar. My intention here is to get us thinking bout how we could use
  this page to communicate who the content is for.