Fix broken links, rework use of MDX partials
While checking the daily 404 logs, I noticed a few unusual lines that
look like this:

```
/docs/flexsave/csp/%5Bobject%20Module%5D
/docs/flexsave/sign-up/gcp/%5Bobject%20Module%5D
```

This told me that there was a JavaScript bug somewhere because when
urldecoded, those URLs look like this:

```
/docs/flexsave/csp/[object Module]
/docs/flexsave/sign-up/gcp/[object Module]
```

JavaScript renders modules as `[object Module]` when typecast to a
string, so some piece of JavaScript code responsible for constructing a
URL was broken somewhere.

Initially, I thought the source of the error might be in the CMP because
I don't recall writing any URL constructors for the Flexsave docs.

Upon further investigation, I realized the source of the error was me
trying to pass modules like `require('../index.mdx')` as parameters to
[MDX partials][mdx-partials], e.g.,:

```jsx
import Preamble from '../_aws-preamble.mdx';

<Preamble indexHref={require('../index.mdx')} />
```

This sort of syntax does [in other contexts](mdx-require), e.g.,:

```jsx
<a href={require('static/files/note.docx')}>Download this document</a>
```

But for now, at least, it seems that I do not understand how to pass
URLs to partials without bypassing the standard Docusaurus `require`
machinery if it is possible at all.

Accordingly, I have refactored my use of partials to eliminate any
partial that requires the a URL as a parameter.

[mdx-partials]: https://docusaurus.io/docs/markdown-features/react#importing-markdown
[mdx-require]: https://docusaurus.io/docs/static-assets#in-markdown