Update logos to use PNGs instead of SVGs
Brad noticed that his browser was rendering the scaled down version of
the PNGs in a way that resulted in a pixelated image.

To fix this problem by asserting pixel perfect control, I have rendered
out both versions of the logo (header and footer) as PNGs a x2
resolution for retina displays.

Now that we are no longer using any of the `brand` assets, I have
deleted them. We can refer to the masters if we need to.