Add a link to the product Changelog to footer
I noticed we are getting some referral traffic from the Changelog, and
being honest, I had kind of forgotten it existed.

We definitely need to promote it more heavily within the Help Center.

This change is a good first start. We now list it above "Feature
requests" under the "Support and consulting" header in the footer.