Update docs on creating support tickets
This PR addresses CMP-2498.

I thoroughly reviewed our docs for raising support tickets and found no
major inaccuracies. However, With Aleks's recent work to help improve
how our customers interact with Severity levels, I saw the opportunity
to carry over some of that effort into the docs.

While I was in the file, I also brought the doc up-to-date with our
newer style guide.

---

If this PR is accepted, I will make a follow-up PR to update the text in
the CMP UI to match my rewording of the response hours sentences.